requirements.txt
Django==5.2.6
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
requests
apscheduler==3.10.4
django-apscheduler==0.7.0
cryptography

